pluggable-architecture/git-data-fetcher/requirements.txt
pytest
pytest-vcr
python-dateutil
requests
